+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to
 - 🩹(frontend) improve reaction toolbar centering with dynamic positioning
 - 🚀 (paas) remove buildpack requirements.txt to use the new uv.lock #1349
 - ✨(backend) allow room configuration and access level via external api #1260
+- ♻️(backend) prefix Swagger routes with /api
 
 ### Fixed
 - 🩹(backend) fix swagger and redoc documentation URLs